+++ SimulationCode/src/Step.h
@@ -4,18 +4,18 @@
 #include "RandomGenerators.h"
 #include <thrust/host_vector.h>
 
-typedef thrust::host_vector<float> tf3;
+typedef thrust::host_vector<double> tf3;
 
 class Step {
 protected:
-  const float G = 6.674 * (1e-11);
-  float oldMomentumX = 0, oldMomentumY = 0, oldMomentumZ = 0;
+  const double G = 6.674 * (1e-11);
+  double oldMomentumX = 0, oldMomentumY = 0, oldMomentumZ = 0;
   RandomGenerators *randomGenerator;
   unsigned numberOfBodies;
 
 public:
   virtual ~Step() {}
-  virtual void compute(tf3 &positions, float dt) = 0;
+  virtual void compute(tf3 &positions, double dt) = 0;
 };
 
 #endif
